pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prajna"
version = "1.1.0"
description = "Predictive Reasoning Architecture for Joint Network-wide Anomalics — Full-Lifecycle Spacecraft Health Intelligence Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Karthik", email = "idikudakarthik55@gmail.com"}]

dependencies = [
    "torch>=2.1.0",
    "torch-geometric>=2.4.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "h5py>=3.9.0",
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "flask>=3.0.0",
    "chromadb>=0.4.0",
    "kagglehub>=0.3.0",
    "tqdm>=4.66.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "plotly>=5.18.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
prajna = "prajna.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["prajna*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
